added recycler view

+++ app/src/main/java/com/example/weatherapp/model/pojo/currentcondition/Metric.java
@@ -7,7 +7,7 @@ public class Metric {
 
     @SerializedName("Value")
     @Expose
-    private Integer value;
+    private Float value;
     @SerializedName("Unit")
     @Expose
     private String unit;
@@ -18,11 +18,11 @@ public class Metric {
     public Metric() {
     }
 
-    public Integer getValue() {
+    public Float getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Float value) {
         this.value = value;
     }
